+++ src/lib/evaluateTailwindFunctions.js
@@ -6,6 +6,7 @@ import { normalizeScreens } from '../util/normalizeScreens'
 import buildMediaQuery from '../util/buildMediaQuery'
 import { toPath } from '../util/toPath'
 import { withAlphaValue } from '../util/withAlphaVariable'
+import { parseColorFormat } from '../util/pluginUtils'
 
 function isObject(input) {
   return typeof input === 'object' && input !== null
@@ -181,6 +182,7 @@ export default function ({ tailwindConfig: config }) {
       }
 
       if (alpha !== undefined) {
+        value = parseColorFormat(value)
         value = withAlphaValue(value, alpha, value)
       }